+++ .changeset/plenty-turkeys-grab.md
@@ -1,0 +1,5 @@
+---
+"studiocms": patch
+---
+
+Remove Deprecated SDK functions and refactor SDK to make it easier to modify/read
