+++ packages/studiocms/src/sdk/consts.ts
@@ -1,0 +1,6 @@
+export const SiteConfigMapID: string = '__StudioCMS_Site_Config';
+export const VersionMapID: string = '__StudioCMS_Latest_Version';
+export const FolderTreeMapID: string = '__StudioCMS_Folder_Tree';
+export const PageFolderTreeMapID: string = '__StudioCMS_Page_Folder_Tree';
+export const FolderListMapID: string = '__StudioCMS_Folder_List';
+export const StudioCMSPkgId: string = 'studiocms';
