+++ packages/studiocms/src/sdk/sdkCore.ts
@@ -1,18 +1,15 @@
 import { and, asc, desc, eq } from 'astro:db';
-import { config, sdk as sdkConfig } from 'studiocms:config';
 import { createTwoFilesPatch } from 'diff';
 import { type Diff2HtmlConfig, html } from 'diff2html';
-import { Effect } from 'effect';
+import { Effect, Layer } from 'effect';
 import {
 	CMSNotificationSettingsId,
 	CMSSiteConfigId,
 	GhostUserDefaults,
 	NotificationSettingsDefaults,
-	versionCacheLifetime,
 } from '../consts.js';
 import {
 	AstroDB,
-	GetVersionFromNPM,
 	SDKCore_Collectors,
 	SDKCore_FolderTree,
 	SDKCore_Generators,
@@ -20,17 +17,19 @@ import {
 	SDKCore_Users,
 } from './effect/index.js';
 import { SDKCoreError, StudioCMS_SDK_Error } from './errors.js';
+import { SDKCore_CLEAR } from './modules/clear.js';
+import { SDKCore_DELETE } from './modules/delete.js';
+import { SDKCore_GET } from './modules/get.js';
+import { SDKCore_POST } from './modules/post.js';
+import { SDKCore_REST_API } from './modules/rest_api.js';
+import { SDKCore_UPDATE } from './modules/update.js';
 import {
-	tsAPIKeys,
 	tsDiffTracking,
 	tsEmailVerificationTokens,
 	tsNotificationSettings,
 	tsOAuthAccounts,
 	tsPageContent,
 	tsPageData,
-	tsPageDataCategories,
-	tsPageDataTags,
-	tsPageFolderStructure,
 	tsPermissions,
 	tsSessionTable,
 	tsSiteConfig,
@@ -38,59 +37,28 @@ import {
 	tsUsers,
 } from './tables.js';
 import type {
-	BaseCacheObject,
-	CombinedInsertContent,
+	CacheMap,
 	CombinedPageData,
 	CombinedUserData,
 	FolderListCacheObject,
-	FolderListItem,
 	FolderNode,
 	FolderTreeCacheObject,
 	MetaOnlyPageData,
-	MetaOnlyPageDataCacheObject,
-	MultiPageInsert,
 	PageDataCacheObject,
-	PageDataCacheReturnType,
-	PaginateInput,
-	SiteConfig,
 	SiteConfigCacheObject,
 	VersionCacheObject,
-	addDatabaseEntryInsertPage,
-	tsDiffTrackingInsert,
-	tsDiffTrackingSelect,
 	tsNotificationSettingsInsert,
 	tsOAuthAccountsSelect,
-	tsPageContentInsert,
-	tsPageContentSelect,
-	tsPageDataCategoriesInsert,
-	tsPageDataCategoriesSelect,
-	tsPageDataInsert,
 	tsPageDataSelect,
-	tsPageDataTagsInsert,
-	tsPageDataTagsSelect,
-	tsPageFolderInsert,
-	tsPageFolderSelect,
-	tsPermissionsInsert,
 	tsPermissionsSelect,
 	tsSessionTableInsert,
-	tsSiteConfigInsert,
 	tsSiteConfigSelect,
 	tsUserResetTokensSelect,
 	tsUsersInsert,
 	tsUsersSelect,
 	tsUsersUpdate,
 } from './types/index.js';
-
-const cacheConfig = sdkConfig.cacheConfig;
-
-const SiteConfigMapID: string = '__StudioCMS_Site_Config';
-const VersionMapID: string = '__StudioCMS_Latest_Version';
-const FolderTreeMapID: string = '__StudioCMS_Folder_Tree';
-const PageFolderTreeMapID: string = '__StudioCMS_Page_Folder_Tree';
-const FolderListMapID: string = '__StudioCMS_Folder_List';
-const StudioCMSPkgId: string = 'studiocms';
-
-type CacheMap<K, V> = ReadonlyMap<K, V> & Map<K, V>;
+import { CacheContext, _ClearUnknownError, _clearLibSQLError } from './utils.js';
 
 const pages: CacheMap<string, PageDataCacheObject> = new Map<string, PageDataCacheObject>();
 const siteConfig: CacheMap<string, SiteConfigCacheObject> = new Map<
@@ -111,649 +79,62 @@ const FolderList: CacheMap<string, FolderListCacheObject> = new Map<
 	FolderListCacheObject
 >();
 
-const isCacheEnabled = Effect.try(() => cacheConfig.enabled);
-
-const _ClearUnknownError = (id: string, cause: unknown) =>
-	Effect.fail(
-		new SDKCoreError({
-			type: 'UNKNOWN',
-			cause: new StudioCMS_SDK_Error(`${id} Error: ${cause}`),
-		})
-	);
-
-const _clearLibSQLError = (id: string, cause: unknown) =>
-	Effect.fail(
-		new SDKCoreError({
-			type: 'LibSQLDatabaseError',
-			cause: new StudioCMS_SDK_Error(`${id} Error: ${cause}`),
-		})
-	);
-
-function folderTreeReturn(data: FolderNode[]): FolderTreeCacheObject {
-	return {
-		data,
-		lastCacheUpdate: new Date(),
-	};
-}
-
-function folderListReturn(data: FolderListItem[]): FolderListCacheObject {
-	return {
-		data,
-		lastCacheUpdate: new Date(),
-	};
-}
-
-function pageDataReturn(data: CombinedPageData): PageDataCacheObject {
-	return {
-		data,
-		lastCacheUpdate: new Date(),
-	};
-}
-
-function siteConfigReturn(siteConfig: SiteConfig): SiteConfigCacheObject {
-	return {
-		data: siteConfig,
-		lastCacheUpdate: new Date(),
-	};
-}
-
-function versionReturn(version: string): VersionCacheObject {
-	return {
-		version,
-		lastCacheUpdate: new Date(),
-	};
-}
-
-function convertCombinedPageDataToMetaOnly<T extends PageDataCacheObject[] | PageDataCacheObject>(
-	data: T
-): PageDataCacheReturnType<T> {
-	if (Array.isArray(data)) {
-		return data.map(({ lastCacheUpdate, data: { defaultContent, multiLangContent, ...data } }) => ({
-			lastCacheUpdate,
-			data,
-		})) as PageDataCacheReturnType<T>;
-	}
-	const {
-		lastCacheUpdate,
-		data: { defaultContent, multiLangContent, ...metaOnlyData },
-	} = data;
-	return {
-		lastCacheUpdate,
-		data: metaOnlyData,
-	} as PageDataCacheReturnType<T>;
-}
-
-function isCacheExpired(entry: BaseCacheObject, lifetime = cacheConfig.lifetime): boolean {
-	return new Date().getTime() - entry.lastCacheUpdate.getTime() > lifetime;
-}
-
-function filterPagesByDraftAndIndex(
-	pages: tsPageDataSelect[],
-	includeDrafts: boolean,
-	hideDefaultIndex: boolean
-): tsPageDataSelect[] {
-	return pages.filter(
-		({ draft, slug }) =>
-			(includeDrafts || draft === false || draft === null) &&
-			(!hideDefaultIndex || slug !== 'index')
-	);
-}
-
 export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', {
+	dependencies: [
+		SDKCore_FolderTree.Default,
+		SDKCore_Generators.Default,
+		SDKCore_Parsers.Default,
+		SDKCore_Users.Default,
+		SDKCore_Collectors.Default,
+		SDKCore_CLEAR.Default,
+		SDKCore_DELETE.Default,
+		SDKCore_UPDATE.Default,
+		SDKCore_POST.Default,
+		SDKCore_REST_API.Default,
+		SDKCore_GET.Default,
+		AstroDB.Default
+	],
 	effect: Effect.gen(function* () {
 		// Get Services
-		const dbService = yield* AstroDB;
-		const folderTreeService = yield* SDKCore_FolderTree;
-		const generatorService = yield* SDKCore_Generators;
-		const parseService = yield* SDKCore_Parsers;
-		const userService = yield* SDKCore_Users;
-		const collectorService = yield* SDKCore_Collectors;
-		const getVersionFromNPM = yield* GetVersionFromNPM;
+		const [
+			{
+				getFullPath,
+				findNodeByPath,
+				findNodesAlongPath,
+				findNodesAlongPathToId,
+				findNodeById,
+				addPageToFolderTree,
+				buildFolderTree,
+				getAvailableFolders,
+			},
+			{ generateRandomIDNumber, generateRandomPassword, generateToken, testToken },
+			{ parseIdNumberArray, parseIdStringArray, fixDiff },
+			{ combineRanks, verifyRank, clearUserReferences },
+			{ collectCategories, collectTags, collectPageData, collectUserData },
+			dbService,
+			CLEAR,
+			DELETE,
+			UPDATE,
+			POST,
+			REST_API,
+			GET,
+		] = yield* Effect.all([
+			SDKCore_FolderTree,
+			SDKCore_Generators,
+			SDKCore_Parsers,
+			SDKCore_Users,
+			SDKCore_Collectors,
+			AstroDB,
+			SDKCore_CLEAR,
+			SDKCore_DELETE,
+			SDKCore_UPDATE,
+			SDKCore_POST,
+			SDKCore_REST_API,
+			SDKCore_GET,
+		]);
 
 		// Breakout service functions that need to be returned in this.
 		const { db } = dbService;
-
-		const {
-			getFullPath,
-			findNodeByPath,
-			findNodesAlongPath,
-			findNodesAlongPathToId,
-			findNodeById,
-			addPageToFolderTree,
-			buildFolderTree,
-			getAvailableFolders,
-		} = folderTreeService;
-
-		const { generateRandomIDNumber, generateRandomPassword, generateToken, testToken } =
-			generatorService;
-
-		const { parseIdNumberArray, parseIdStringArray } = parseService;
-
-		const { combineRanks, verifyRank, clearUserReferences } = userService;
-
-		const { collectCategories, collectTags, collectPageData, collectUserData } = collectorService;
-
-		function _getPackagesPages(
-			packageName: string,
-			tree?: FolderNode[]
-		): Effect.Effect<CombinedPageData[], SDKCoreError, never>;
-		function _getPackagesPages(
-			packageName: string,
-			tree?: FolderNode[],
-			metaOnly?: boolean
-		): Effect.Effect<MetaOnlyPageData[], SDKCoreError, never>;
-
-		function _getPackagesPages(packageName: string, tree?: FolderNode[], metaOnly = false) {
-			return Effect.gen(function* () {
-				const pagesRaw = yield* dbService.execute((db) =>
-					db.select().from(tsPageData).where(eq(tsPageData.package, packageName))
-				);
-
-				const folders = tree || (yield* buildFolderTree);
-
-				const pages = [];
-
-				for (const page of pagesRaw) {
-					const data = yield* collectPageData(page, folders, metaOnly);
-					pages.push(data);
-				}
-
-				return pages as MetaOnlyPageData[] | CombinedPageData[];
-			}).pipe(
-				Effect.catchTags({
-					'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-						_clearLibSQLError('GET.packagePages', cause),
-				})
-			);
-		}
-
-		function _getPageById(id: string): Effect.Effect<PageDataCacheObject, SDKCoreError, never>;
-		function _getPageById(
-			id: string,
-			metaOnly?: boolean
-		): Effect.Effect<MetaOnlyPageDataCacheObject, SDKCoreError, never>;
-
-		function _getPageById(id: string, metaOnly = false) {
-			const getPage = (id: string, tree?: FolderNode[]) =>
-				Effect.gen(function* () {
-					const page = yield* dbService.execute((db) =>
-						db.select().from(tsPageData).where(eq(tsPageData.id, id)).get()
-					);
-
-					if (!page) return undefined;
-
-					const folders = tree || (yield* buildFolderTree);
-
-					return yield* collectPageData(page, folders);
-				});
-
-			return Effect.gen(function* () {
-				const status = yield* isCacheEnabled;
-
-				if (!status) {
-					const page = yield* getPage(id);
-
-					if (!page) {
-						return yield* Effect.fail(
-							new SDKCoreError({
-								type: 'UNKNOWN',
-								cause: new StudioCMS_SDK_Error('Page not found in Database'),
-							})
-						);
-					}
-
-					const pageData = pageDataReturn(page);
-
-					return metaOnly ? convertCombinedPageDataToMetaOnly(pageData) : pageData;
-				}
-
-				const { data: tree } = yield* GET.folderTree();
-
-				const cachedPage = pages.get(id);
-
-				if (!cachedPage || isCacheExpired(cachedPage)) {
-					const page = yield* getPage(id, tree);
-
-					if (!page) {
-						return yield* Effect.fail(
-							new SDKCoreError({
-								type: 'UNKNOWN',
-								cause: new StudioCMS_SDK_Error('Page not found in Database'),
-							})
-						);
-					}
-
-					const returnPage = pageDataReturn(page);
-
-					pages.set(id, returnPage);
-
-					return metaOnly ? convertCombinedPageDataToMetaOnly(returnPage) : returnPage;
-				}
-
-				// Return the cached page
-				return metaOnly ? convertCombinedPageDataToMetaOnly(cachedPage) : cachedPage;
-			}).pipe(
-				Effect.catchTags({
-					UnknownException: (cause) => _clearLibSQLError('GET.page.byId', cause),
-					'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-						_clearLibSQLError('GET.page.byId', cause),
-				})
-			);
-		}
-
-		function _getPageBySlug(slug: string): Effect.Effect<PageDataCacheObject, SDKCoreError, never>;
-		function _getPageBySlug(
-			slug: string,
-			metaOnly?: boolean
-		): Effect.Effect<MetaOnlyPageDataCacheObject, SDKCoreError, never>;
-
-		function _getPageBySlug(slug: string, metaOnly = false) {
-			const getPage = (iSlug: string, tree?: FolderNode[]) =>
-				Effect.gen(function* () {
-					const page = yield* dbService.execute((db) =>
-						db.select().from(tsPageData).where(eq(tsPageData.slug, iSlug)).get()
-					);
-
-					if (!page) return undefined;
-
-					const folders = tree || (yield* buildFolderTree);
-
-					return yield* collectPageData(page, folders);
-				});
-
-			return Effect.gen(function* () {
-				const status = yield* isCacheEnabled;
-
-				if (!status) {
-					const page = yield* getPage(slug);
-
-					if (!page) {
-						return yield* Effect.fail(
-							new SDKCoreError({
-								type: 'UNKNOWN',
-								cause: new StudioCMS_SDK_Error('Page not found in Database'),
-							})
-						);
-					}
-
-					const pageData = pageDataReturn(page);
-
-					return metaOnly ? convertCombinedPageDataToMetaOnly(pageData) : pageData;
-				}
-
-				const { data: tree } = yield* GET.folderTree();
-
-				// Retrieve the cached page
-				const cachedPage = Array.from(pages.values()).find((page) => page.data.slug === slug);
-
-				if (!cachedPage || isCacheExpired(cachedPage)) {
-					const page = yield* getPage(slug, tree);
-
-					if (!page) {
-						return yield* Effect.fail(
-							new SDKCoreError({
-								type: 'UNKNOWN',
-								cause: new StudioCMS_SDK_Error('Page not found in Database'),
-							})
-						);
-					}
-
-					const pageData = pageDataReturn(page);
-
-					return metaOnly ? convertCombinedPageDataToMetaOnly(pageData) : pageData;
-				}
-
-				// Return the cached page
-				return metaOnly ? convertCombinedPageDataToMetaOnly(cachedPage) : cachedPage;
-			}).pipe(
-				Effect.catchTags({
-					UnknownException: (cause) => _clearLibSQLError('GET.page.bySlug', cause),
-					'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-						_clearLibSQLError('GET.page.bySlug', cause),
-				})
-			);
-		}
-
-		function _getAllPages(
-			includeDrafts?: boolean,
-			hideDefaultIndex?: boolean,
-			metaOnly?: false,
-			paginate?: PaginateInput
-		): Effect.Effect<PageDataCacheObject[], SDKCoreError, never>;
-		function _getAllPages(
-			includeDrafts?: boolean,
-			hideDefaultIndex?: boolean,
-			metaOnly?: true,
-			paginate?: PaginateInput
-		): Effect.Effect<MetaOnlyPageDataCacheObject[], SDKCoreError, never>;
-
-		function _getAllPages(
-			includeDrafts = false,
-			hideDefaultIndex = false,
-			metaOnly = false,
-			paginate?: PaginateInput
-		) {
-			const getPages = (
-				includeDrafts = false,
-				hideDefaultIndex = false,
-				tree?: FolderNode[],
-				paginate?: PaginateInput
-			) =>
-				Effect.gen(function* () {
-					if (paginate) {
-						if (paginate.limit < 0 || paginate.offset < 0) {
-							return yield* Effect.fail(
-								new SDKCoreError({
-									type: 'UNKNOWN',
-									cause: new StudioCMS_SDK_Error(
-										'Pagination limit and offset must be non-negative values'
-									),
-								})
-							);
-						}
-						if (paginate.limit === 0) {
-							// Either throw an error or set a default value
-							paginate.limit = 10; // Default value
-						}
-					}
-
-					const pagesRaw = paginate
-						? yield* dbService.execute((db) =>
-								db
-									.select()
-									.from(tsPageData)
-									.orderBy(asc(tsPageData.title))
-									.limit(paginate.limit)
-									.offset(paginate.offset)
-							)
-						: yield* dbService.execute((db) =>
-								db.select().from(tsPageData).orderBy(asc(tsPageData.title))
-							);
-
-					const pagesFiltered = filterPagesByDraftAndIndex(
-						pagesRaw,
-						includeDrafts,
-						hideDefaultIndex
-					);
-
-					const folders = tree || (yield* buildFolderTree);
-
-					const pages = [];
-
-					for (const page of pagesFiltered) {
-						const data = yield* collectPageData(page, folders);
-						pages.push(data);
-					}
-
-					return pages;
-				});
-
-			return Effect.gen(function* () {
-				const status = yield* isCacheEnabled;
-
-				if (paginate) {
-					if (paginate.limit < 0 || paginate.offset < 0) {
-						return yield* Effect.fail(
-							new SDKCoreError({
-								type: 'UNKNOWN',
-								cause: new StudioCMS_SDK_Error(
-									'Pagination limit and offset must be non-negative values'
-								),
-							})
-						);
-					}
-					if (paginate.limit === 0) {
-						// Either throw an error or set a default value
-						paginate.limit = 10; // Default value
-					}
-				}
-
-				if (!status) {
-					const dbPages = yield* getPages(includeDrafts, hideDefaultIndex, undefined, paginate);
-
-					const data = dbPages.map((page) => pageDataReturn(page));
-					return metaOnly ? convertCombinedPageDataToMetaOnly(data) : data;
-				}
-
-				const { data: tree } = yield* GET.folderTree();
-
-				if (pages.size === 0) {
-					const newData = yield* getPages(includeDrafts, hideDefaultIndex, tree);
-
-					// Loop through the updated data and store it in the cache
-					for (const data of newData) {
-						pages.set(data.id, pageDataReturn(data));
-					}
-
-					const data = newData.map((data) => pageDataReturn(data));
-
-					if (paginate) {
-						const paginatedData = data
-							.sort((a, b) => a.data.title.localeCompare(b.data.title))
-							.slice(paginate.offset, paginate.offset + paginate.limit);
-						return metaOnly ? convertCombinedPageDataToMetaOnly(paginatedData) : paginatedData;
-					}
-
-					// Transform and return the data
-					return metaOnly ? convertCombinedPageDataToMetaOnly(data) : data;
-				}
-
-				const cacheMap = Array.from(pages.values());
-
-				for (const item of cacheMap) {
-					if (isCacheExpired(item)) {
-						const { data: updatedData } = yield* GET.page.byId(item.data.id);
-
-						pages.set(updatedData.id, pageDataReturn(updatedData));
-					}
-				}
-
-				const data = Array.from(pages.values());
-
-				if (paginate) {
-					const paginatedData = data
-						.sort((a, b) => a.data.title.localeCompare(b.data.title))
-						.slice(paginate.offset, paginate.offset + paginate.limit);
-					return metaOnly ? convertCombinedPageDataToMetaOnly(paginatedData) : paginatedData;
-				}
-
-				return metaOnly ? convertCombinedPageDataToMetaOnly(data) : data;
-			}).pipe(
-				Effect.catchTags({
-					UnknownException: (cause) => _clearLibSQLError('GET.pages', cause),
-					'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-						_clearLibSQLError('GET.pages', cause),
-				})
-			);
-		}
-
-		function _folderPages(
-			id: string,
-			includeDrafts?: boolean,
-			hideDefaultIndex?: boolean,
-			metaOnly?: false,
-			paginate?: PaginateInput
-		): Effect.Effect<PageDataCacheObject[], SDKCoreError, never>;
-		function _folderPages(
-			id: string,
-			includeDrafts?: boolean,
-			hideDefaultIndex?: boolean,
-			metaOnly?: true,
-			paginate?: PaginateInput
-		): Effect.Effect<MetaOnlyPageDataCacheObject[], SDKCoreError, never>;
-
-		function _folderPages(
-			id: string,
-			includeDrafts = false,
-			hideDefaultIndex = false,
-			metaOnly = false,
-			paginate?: PaginateInput
-		) {
-			const getAllPages = (
-				includeDrafts = false,
-				hideDefaultIndex = false,
-				tree?: FolderNode[],
-				paginate?: PaginateInput
-			) =>
-				Effect.gen(function* () {
-					if (paginate) {
-						if (paginate.limit < 0 || paginate.offset < 0) {
-							return yield* Effect.fail(
-								new SDKCoreError({
-									type: 'UNKNOWN',
-									cause: new StudioCMS_SDK_Error(
-										'Pagination limit and offset must be non-negative values'
-									),
-								})
-							);
-						}
-						if (paginate.limit === 0) {
-							// Either throw an error or set a default value
-							paginate.limit = 10; // Default value
-						}
-					}
-
-					const pagesRaw = paginate
-						? yield* dbService.execute((db) =>
-								db
-									.select()
-									.from(tsPageData)
-									.orderBy(asc(tsPageData.title))
-									.limit(paginate.limit)
-									.offset(paginate.offset)
-							)
-						: yield* dbService.execute((db) =>
-								db.select().from(tsPageData).orderBy(asc(tsPageData.title))
-							);
-
-					const pagesFiltered = filterPagesByDraftAndIndex(
-						pagesRaw,
-						includeDrafts,
-						hideDefaultIndex
-					);
-
-					const folders = tree || (yield* buildFolderTree);
-
-					const pages = [];
-
-					for (const page of pagesFiltered) {
-						const data = yield* collectPageData(page, folders);
-						pages.push(data);
-					}
-
-					return pages;
-				});
-			const getPages = (
-				id: string,
-				includeDrafts = false,
-				hideDefaultIndex = false,
-				tree?: FolderNode[]
-			) =>
-				Effect.gen(function* () {
-					const pagesRaw = yield* dbService.execute((db) =>
-						db
-							.select()
-							.from(tsPageData)
-							.where(eq(tsPageData.parentFolder, id))
-							.orderBy(asc(tsPageData.title))
-					);
-
-					const pagesFiltered = filterPagesByDraftAndIndex(
-						pagesRaw,
-						includeDrafts,
-						hideDefaultIndex
-					);
-
-					const folders = tree || (yield* buildFolderTree);
-
-					const pages = [];
-
-					for (const page of pagesFiltered) {
-						pages.push(yield* collectPageData(page, folders));
-					}
-
-					return pages;
-				});
-
-			return Effect.gen(function* () {
-				const status = yield* isCacheEnabled;
-
-				if (paginate) {
-					if (paginate.limit < 0 || paginate.offset < 0) {
-						return yield* Effect.fail(
-							new SDKCoreError({
-								type: 'UNKNOWN',
-								cause: new StudioCMS_SDK_Error(
-									'Pagination limit and offset must be non-negative values'
-								),
-							})
-						);
-					}
-					if (paginate.limit === 0) {
-						// Either throw an error or set a default value
-						paginate.limit = 10; // Default value
-					}
-				}
-
-				if (!status) {
-					const dbPages = yield* getPages(id, includeDrafts, hideDefaultIndex);
-					const data = dbPages.map((page) => pageDataReturn(page));
-					return metaOnly ? convertCombinedPageDataToMetaOnly(data) : data;
-				}
-
-				const { data: tree } = yield* GET.folderTree();
-
-				if (pages.size === 0) {
-					const updatedData = yield* getAllPages(includeDrafts, hideDefaultIndex, tree);
-
-					for (const data of updatedData) {
-						pages.set(data.id, pageDataReturn(data));
-					}
-
-					const data = updatedData
-						.filter(({ parentFolder }) => parentFolder === id)
-						.map((data) => pageDataReturn(data));
-
-					if (paginate) {
-						const paginatedData = data
-							.sort((a, b) => a.data.title.localeCompare(b.data.title))
-							.slice(paginate.offset, paginate.offset + paginate.limit);
-						return metaOnly ? convertCombinedPageDataToMetaOnly(paginatedData) : paginatedData;
-					}
-
-					// Transform and return the data
-					return metaOnly ? convertCombinedPageDataToMetaOnly(data) : data;
-				}
-
-				const cacheMap = Array.from(pages.values());
-
-				for (const item of cacheMap) {
-					if (isCacheExpired(item)) {
-						const { data: updatedData } = yield* GET.page.byId(item.data.id);
-
-						pages.set(updatedData.id, pageDataReturn(updatedData));
-					}
-				}
-
-				const data = Array.from(pages.values()).filter(
-					({ data: { parentFolder } }) => parentFolder === id
-				);
-
-				if (paginate) {
-					const paginatedData = data
-						.sort((a, b) => a.data.title.localeCompare(b.data.title))
-						.slice(paginate.offset, paginate.offset + paginate.limit);
-					return metaOnly ? convertCombinedPageDataToMetaOnly(paginatedData) : paginatedData;
-				}
-				return metaOnly ? convertCombinedPageDataToMetaOnly(data) : data;
-			}).pipe(
-				Effect.catchTags({
-					UnknownException: (cause) => _clearLibSQLError('GET.pages', cause),
-					'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-						_clearLibSQLError('GET.pages', cause),
-				})
-			);
-		}
 
 		const resetTokenBucket = {
 			new: (userId: string): Effect.Effect<tsUserResetTokensSelect, SDKCoreError, never> =>
@@ -904,7 +285,7 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 							.get()
 					);
 
-					return yield* parseService.fixDiff(inputted);
+					return yield* fixDiff(inputted);
 				}).pipe(
 					Effect.catchTags({
 						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
@@ -944,7 +325,7 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 									.orderBy(desc(tsDiffTracking.timestamp))
 							);
 
-							return yield* parseService.fixDiff(items);
+							return yield* fixDiff(items);
 						}).pipe(
 							Effect.catchTags({
 								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
@@ -970,7 +351,7 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 
 							const split = items.slice(0, count);
 
-							return yield* parseService.fixDiff(split);
+							return yield* fixDiff(split);
 						}).pipe(
 							Effect.catchTags({
 								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
@@ -996,7 +377,7 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 									.orderBy(desc(tsDiffTracking.timestamp))
 							);
 
-							return yield* parseService.fixDiff(items);
+							return yield* fixDiff(items);
 						}).pipe(
 							Effect.catchTags({
 								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
@@ -1022,7 +403,7 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 
 							const split = items.slice(0, count);
 
-							return yield* parseService.fixDiff(split);
+							return yield* fixDiff(split);
 						}).pipe(
 							Effect.catchTags({
 								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
@@ -1043,7 +424,7 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 							db.select().from(tsDiffTracking).where(eq(tsDiffTracking.id, id)).get()
 						);
 						if (!data) return;
-						return yield* parseService.fixDiff(data);
+						return yield* fixDiff(data);
 					}).pipe(
 						Effect.catchTags({
 							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
@@ -1112,7 +493,7 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 						yield* purgeDiff(id);
 					}
 
-					return yield* parseService.fixDiff(diffEntry);
+					return yield* fixDiff(diffEntry);
 				}).pipe(
 					Effect.catchTags({
 						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
@@ -1780,1584 +1161,6 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 			},
 		};
 
-		const DELETE = {
-			/**
-			 * Deletes a page from the database.
-			 *
-			 * @param id - The ID of the page to delete.
-			 * @returns A promise that resolves to a deletion response.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the page.
-			 */
-			page: (id: string) =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db.delete(tsDiffTracking).where(eq(tsDiffTracking.pageId, id))
-					);
-					yield* dbService.execute((db) =>
-						db.delete(tsPageContent).where(eq(tsPageContent.contentId, id))
-					);
-					yield* dbService.execute((db) => db.delete(tsPageData).where(eq(tsPageData.id, id)));
-
-					yield* CLEAR.pages();
-					return {
-						status: 'success',
-						message: `Page with ID ${id} has been deleted successfully`,
-					};
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('DELETE.page', cause),
-					})
-				),
-			/**
-			 * Deletes a page content from the database.
-			 *
-			 * @param id - The ID of the page content to delete.
-			 * @returns A promise that resolves to a deletion response.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the page content.
-			 */
-			pageContent: (id: string) =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db.delete(tsPageContent).where(eq(tsPageContent.contentId, id))
-					);
-					return {
-						status: 'success',
-						message: `Page content with ID ${id} has been deleted successfully`,
-					};
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('DELETE.pageContent', cause),
-					})
-				),
-			/**
-			 * Deletes a page content lang from the database.
-			 *
-			 * @param id - The ID of the page content to delete.
-			 * @param lang - The lang of the page content to delete.
-			 * @returns A promise that resolves to a deletion response.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the page content lang.
-			 */
-			pageContentLang: (id: string, lang: string) =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db
-							.delete(tsPageContent)
-							.where(and(eq(tsPageContent.contentId, id), eq(tsPageContent.contentLang, lang)))
-					);
-					return {
-						status: 'success',
-						message: `Page content with ID ${id} and lang ${lang} has been deleted successfully`,
-					};
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('DELETE.pageContentLang', cause),
-					})
-				),
-			/**
-			 * Deletes a tag from the database.
-			 *
-			 * @param id - The ID of the tag to delete.
-			 * @returns A promise that resolves to a deletion response.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the tag.
-			 */
-			tags: (id: number) =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db.delete(tsPageDataTags).where(eq(tsPageDataTags.id, id))
-					);
-					return {
-						status: 'success',
-						message: `Tag with ID ${id} has been deleted successfully`,
-					};
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('DELETE.tags', cause),
-					})
-				),
-			/**
-			 * Deletes a category from the database.
-			 *
-			 * @param id - The ID of the category to delete.
-			 * @returns A promise that resolves to a deletion response.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the category.
-			 */
-			categories: (id: number) =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db.delete(tsPageDataCategories).where(eq(tsPageDataCategories.id, id))
-					);
-					return {
-						status: 'success',
-						message: `Category with ID ${id} has been deleted successfully`,
-					};
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('DELETE.categories', cause),
-					})
-				),
-			/**
-			 * Deletes a permission from the database.
-			 *
-			 * @param userId - The ID of the user to delete the permission for.
-			 * @returns A promise that resolves to a deletion response.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the permission.
-			 */
-			permissions: (userId: string) =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db.delete(tsPermissions).where(eq(tsPermissions.user, userId))
-					);
-					return {
-						status: 'success',
-						message: `Permissions for user with ID ${userId} have been deleted successfully`,
-					};
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('DELETE.permissions', cause),
-					})
-				),
-			/**
-			 * Deletes a site configuration from the database.
-			 *
-			 * @param id - The ID of the site configuration to delete.
-			 * @returns A promise that resolves to a deletion response.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the site configuration.
-			 */
-			diffTracking: (id: string) =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db.delete(tsDiffTracking).where(eq(tsDiffTracking.id, id))
-					);
-					return {
-						status: 'success',
-						message: `Diff tracking with ID ${id} has been deleted successfully`,
-					};
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('DELETE.diffTracking', cause),
-					})
-				),
-			/**
-			 * Deletes a folder from the database.
-			 *
-			 * @param id - The ID of the folder to delete.
-			 * @returns A promise that resolves to a deletion response.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the folder.
-			 */
-			folder: (id: string) =>
-				Effect.gen(function* () {
-					yield* dbService.execute((db) =>
-						db.delete(tsPageFolderStructure).where(eq(tsPageFolderStructure.id, id))
-					);
-
-					yield* CLEAR.folderList();
-					yield* CLEAR.folderTree();
-
-					yield* UPDATE.folderList;
-					yield* UPDATE.folderTree;
-
-					return {
-						status: 'success',
-						message: `Folder with ID ${id} has been deleted successfully`,
-					};
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('DELETE.folder', cause),
-					})
-				),
-			/**
-			 * Deletes a user from the database.
-			 *
-			 * @param id - The ID of the user to delete.
-			 * @returns A promise that resolves to a deletion response.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while deleting the user.
-			 */
-			user: (id: string) =>
-				Effect.gen(function* () {
-					if (id === GhostUserDefaults.id) {
-						return yield* _clearLibSQLError(
-							'DELETE.user',
-							`User with ID ${id} is an internal user and cannot be deleted.`
-						);
-					}
-
-					const verifyNoReference = yield* clearUserReferences(id);
-
-					if (!verifyNoReference) {
-						return yield* _clearLibSQLError(
-							'DELETE.user',
-							`There was an issue deleting User with ID ${id}. Please manually remove all references before deleting the user. Or try again.`
-						);
-					}
-
-					yield* dbService.execute((db) => db.delete(tsUsers).where(eq(tsUsers.id, id)));
-
-					return {
-						status: 'success',
-						message: `User with ID ${id} has been deleted successfully`,
-					};
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('DELETE.user', cause),
-					})
-				),
-		};
-
-		const CLEAR = {
-			page: {
-				byId: (id: string) =>
-					Effect.gen(function* () {
-						const status = yield* isCacheEnabled;
-						if (!status) return;
-
-						pages.delete(id);
-						return;
-					}).pipe(
-						Effect.catchTags({
-							UnknownException: (cause) => _ClearUnknownError('CLEAR.page.byId', cause),
-						})
-					),
-				bySlug: (slug: string) =>
-					Effect.gen(function* () {
-						const status = yield* isCacheEnabled;
-						if (!status) return;
-
-						const keyIndex: string[] = [];
-
-						for (const [key, cachedObject] of pages) {
-							if (cachedObject.data.slug === slug) {
-								keyIndex.push(key);
-							}
-						}
-
-						for (const key of keyIndex) {
-							pages.delete(key);
-						}
-
-						return;
-					}).pipe(
-						Effect.catchTags({
-							UnknownException: (cause) => _ClearUnknownError('CLEAR.page.bySlug', cause),
-						})
-					),
-			},
-			pages: () =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-					if (!status) return;
-
-					pages.clear();
-					pageFolderTree.clear();
-					folderTree.clear();
-					FolderList.clear();
-					return;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _ClearUnknownError('CLEAR.pages', cause),
-					})
-				),
-			latestVersion: () =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-					if (!status) return;
-
-					version.clear();
-					return;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _ClearUnknownError('CLEAR.latestVersion', cause),
-					})
-				),
-			folderTree: () =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-					if (!status) return;
-
-					folderTree.clear();
-					pageFolderTree.clear();
-					return;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _ClearUnknownError('CLEAR.folderTree', cause),
-					})
-				),
-			folderList: () =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-					if (!status) return;
-
-					FolderList.clear();
-					return;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _ClearUnknownError('CLEAR.folderList', cause),
-					})
-				),
-		};
-
-		const REST_API = {
-			tokens: {
-				get: dbService.makeQuery((ex, userId: string) =>
-					ex((db) => db.select().from(tsAPIKeys).where(eq(tsAPIKeys.userId, userId))).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('REST_API.tokens.get', cause),
-						})
-					)
-				),
-				new: (userId: string, description: string) =>
-					Effect.gen(function* () {
-						const key = yield* generateToken(userId, true);
-
-						return yield* dbService.execute((db) =>
-							db
-								.insert(tsAPIKeys)
-								.values({
-									// @ts-expect-error Drizzle... removed this from the type?
-									id: crypto.randomUUID(),
-									creationDate: new Date(),
-									userId,
-									key,
-									description,
-								})
-								.returning()
-								.get()
-						);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('REST_API.tokens.new', cause),
-						})
-					),
-				delete: (userId: string, tokenId: string) =>
-					dbService
-						.execute((db) =>
-							db
-								.delete(tsAPIKeys)
-								.where(and(eq(tsAPIKeys.userId, userId), eq(tsAPIKeys.id, tokenId)))
-						)
-						.pipe(
-							Effect.catchTags({
-								'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-									_clearLibSQLError('REST_API.tokens.delete', cause),
-							})
-						),
-				verify: (key: string) =>
-					Effect.gen(function* () {
-						const apiKey = yield* dbService.execute((db) =>
-							db.select().from(tsAPIKeys).where(eq(tsAPIKeys.key, key)).get()
-						);
-
-						if (!apiKey) return false;
-
-						const keyRank = yield* dbService.execute((db) =>
-							db.select().from(tsPermissions).where(eq(tsPermissions.user, apiKey.userId)).get()
-						);
-
-						if (!keyRank) return false;
-
-						return {
-							userId: apiKey.userId,
-							key: apiKey.key,
-							rank: keyRank.rank,
-						};
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('REST_API.tokens.verify', cause),
-						})
-					),
-			},
-		};
-
-		const GET = {
-			databaseTable: {
-				users: () => dbService.execute((db) => db.select().from(tsUsers)),
-				oAuthAccounts: () => dbService.execute((db) => db.select().from(tsOAuthAccounts)),
-				sessionTable: () => dbService.execute((db) => db.select().from(tsSessionTable)),
-				permissions: () => dbService.execute((db) => db.select().from(tsPermissions)),
-				pageData: () => dbService.execute((db) => db.select().from(tsPageData)),
-				pageDataTags: () => dbService.execute((db) => db.select().from(tsPageDataTags)),
-				pageDataCategories: () => dbService.execute((db) => db.select().from(tsPageDataCategories)),
-				pageContent: () => dbService.execute((db) => db.select().from(tsPageContent)),
-				siteConfig: () =>
-					dbService.execute((db) =>
-						db.select().from(tsSiteConfig).where(eq(tsSiteConfig.id, CMSSiteConfigId)).get()
-					),
-				diffTracking: () => dbService.execute((db) => db.select().from(tsDiffTracking)),
-				pageFolderStructure: () =>
-					dbService.execute((db) => db.select().from(tsPageFolderStructure)),
-				notificationSettings: () =>
-					dbService.execute((db) =>
-						db
-							.select()
-							.from(tsNotificationSettings)
-							.where(eq(tsNotificationSettings.id, CMSNotificationSettingsId))
-							.get()
-					),
-				emailVerificationTokens: () =>
-					dbService.execute((db) => db.select().from(tsEmailVerificationTokens)),
-			},
-			permissionsLists: {
-				/**
-				 * Retrieves all permissions for users in the database.
-				 *
-				 * @returns A promise that resolves to an array of combined rank data.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the permissions.
-				 */
-				all: () =>
-					Effect.gen(function* () {
-						const currentPermittedUsers = yield* dbService.execute((db) =>
-							db.select().from(tsPermissions)
-						);
-						const existingUsers = yield* dbService.execute((db) => db.select().from(tsUsers));
-
-						const owners = yield* verifyRank(existingUsers, currentPermittedUsers, 'owner');
-
-						const admins = yield* verifyRank(existingUsers, currentPermittedUsers, 'admin');
-
-						const editors = yield* verifyRank(existingUsers, currentPermittedUsers, 'editor');
-
-						const visitors = yield* verifyRank(existingUsers, currentPermittedUsers, 'visitor');
-
-						return [
-							...(yield* combineRanks('owner', owners)),
-							...(yield* combineRanks('admin', admins)),
-							...(yield* combineRanks('editor', editors)),
-							...(yield* combineRanks('visitor', visitors)),
-						];
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('GET.permissionsLists.all', cause),
-						})
-					),
-				/**
-				 * Retrieves all owners in the database.
-				 *
-				 * @returns A promise that resolves to an array of combined rank data.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the owners.
-				 */
-				owners: () =>
-					Effect.gen(function* () {
-						const currentPermittedUsers = yield* dbService.execute((db) =>
-							db.select().from(tsPermissions)
-						);
-						const existingUsers = yield* dbService.execute((db) => db.select().from(tsUsers));
-
-						return yield* verifyRank(existingUsers, currentPermittedUsers, 'owner');
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('GET.permissionsLists.owners', cause),
-						})
-					),
-				/**
-				 * Retrieves all admins in the database.
-				 *
-				 * @returns A promise that resolves to an array of combined rank data.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the admins.
-				 */
-				admins: () =>
-					Effect.gen(function* () {
-						const currentPermittedUsers = yield* dbService.execute((db) =>
-							db.select().from(tsPermissions)
-						);
-						const existingUsers = yield* dbService.execute((db) => db.select().from(tsUsers));
-
-						return yield* verifyRank(existingUsers, currentPermittedUsers, 'admin');
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('GET.permissionsLists.admins', cause),
-						})
-					),
-				/**
-				 * Retrieves all editors in the database.
-				 *
-				 * @returns A promise that resolves to an array of combined rank data.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the editors.
-				 */
-				editors: () =>
-					Effect.gen(function* () {
-						const currentPermittedUsers = yield* dbService.execute((db) =>
-							db.select().from(tsPermissions)
-						);
-						const existingUsers = yield* dbService.execute((db) => db.select().from(tsUsers));
-
-						return yield* verifyRank(existingUsers, currentPermittedUsers, 'editor');
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('GET.permissionsLists.editors', cause),
-						})
-					),
-				/**
-				 * Retrieves all visitors in the database.
-				 *
-				 * @returns A promise that resolves to an array of combined rank data.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the visitors.
-				 */
-				visitors: () =>
-					Effect.gen(function* () {
-						const currentPermittedUsers = yield* dbService.execute((db) =>
-							db.select().from(tsPermissions)
-						);
-						const existingUsers = yield* dbService.execute((db) => db.select().from(tsUsers));
-
-						return yield* verifyRank(existingUsers, currentPermittedUsers, 'visitor');
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('GET.permissionsLists.visitors', cause),
-						})
-					),
-			},
-			users: {
-				/**
-				 * Retrieves all users from the database.
-				 *
-				 * @returns A promise that resolves to an array of combined user data.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the users.
-				 */
-				all: () =>
-					Effect.gen(function* () {
-						const combinedUserData: CombinedUserData[] = [];
-
-						const users = yield* dbService.execute((db) => db.select().from(tsUsers));
-
-						for (const user of users) {
-							if (user.id === GhostUserDefaults.id) {
-								continue;
-							}
-
-							const UserData = yield* collectUserData(user);
-
-							combinedUserData.push(UserData);
-						}
-
-						return combinedUserData;
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('GET.users.all', cause),
-						})
-					),
-				/**
-				 * Retrieves a user by ID.
-				 *
-				 * @param id - The ID of the user to retrieve.
-				 * @returns A promise that resolves to the user data.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the user.
-				 */
-				byId: (id: string) =>
-					Effect.gen(function* () {
-						const user = yield* dbService.execute((db) =>
-							db.select().from(tsUsers).where(eq(tsUsers.id, id)).get()
-						);
-
-						if (!user) return undefined;
-						return yield* collectUserData(user);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('GET.users.byId', cause),
-						})
-					),
-				/**
-				 * Retrieves a user by username.
-				 *
-				 * @param username - The username of the user to retrieve.
-				 * @returns A promise that resolves to the user data.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the user.
-				 */
-				byUsername: (username: string) =>
-					Effect.gen(function* () {
-						const user = yield* dbService.execute((db) =>
-							db.select().from(tsUsers).where(eq(tsUsers.username, username)).get()
-						);
-
-						if (!user) return undefined;
-						return yield* collectUserData(user);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('GET.users.byId', cause),
-						})
-					),
-				/**
-				 * Retrieves a user by email.
-				 *
-				 * @param email - The email of the user to retrieve.
-				 * @returns A promise that resolves to the user data.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while getting the user.
-				 */
-				byEmail: (email: string) =>
-					Effect.gen(function* () {
-						const user = yield* dbService.execute((db) =>
-							db.select().from(tsUsers).where(eq(tsUsers.email, email)).get()
-						);
-
-						if (!user) return undefined;
-						return yield* collectUserData(user);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('GET.users.byId', cause),
-						})
-					),
-			},
-			folder: dbService.makeQuery((ex, id: string) =>
-				ex((db) =>
-					db.select().from(tsPageFolderStructure).where(eq(tsPageFolderStructure.id, id)).get()
-				).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('GET.folder', cause),
-					})
-				)
-			),
-			siteConfig: () =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-
-					if (config.dbStartPage) {
-						return undefined;
-					}
-
-					if (!status) {
-						const newConfig = yield* dbService.execute((db) =>
-							db.select().from(tsSiteConfig).where(eq(tsSiteConfig.id, CMSSiteConfigId)).get()
-						);
-
-						if (!newConfig) {
-							return yield* Effect.fail(
-								new SDKCoreError({
-									type: 'UNKNOWN',
-									cause: new StudioCMS_SDK_Error('Site config not found in database'),
-								})
-							);
-						}
-
-						return siteConfigReturn(newConfig);
-					}
-
-					const currentSiteConfig = siteConfig.get(SiteConfigMapID);
-
-					if (!currentSiteConfig || isCacheExpired(currentSiteConfig)) {
-						const newConfig = yield* dbService.execute((db) =>
-							db.select().from(tsSiteConfig).where(eq(tsSiteConfig.id, CMSSiteConfigId)).get()
-						);
-
-						if (!newConfig) {
-							return yield* Effect.fail(
-								new SDKCoreError({
-									type: 'UNKNOWN',
-									cause: new StudioCMS_SDK_Error('Site config not found in database'),
-								})
-							);
-						}
-
-						const returnConfig = siteConfigReturn(newConfig);
-
-						siteConfig.set(SiteConfigMapID, returnConfig);
-
-						return returnConfig;
-					}
-
-					return currentSiteConfig;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _clearLibSQLError('GET.siteConfig', cause),
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('GET.siteConfig', cause),
-					})
-				),
-			folderTree: () =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-
-					if (!status) {
-						return folderTreeReturn(yield* buildFolderTree);
-					}
-
-					const tree = folderTree.get(FolderTreeMapID);
-
-					if (!tree || isCacheExpired(tree)) {
-						const newFolderTree = yield* buildFolderTree;
-
-						const returnable = folderTreeReturn(newFolderTree);
-
-						folderTree.set(FolderTreeMapID, returnable);
-
-						return returnable;
-					}
-
-					return tree;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _clearLibSQLError('GET.folderTree', cause),
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('GET.folderTree', cause),
-					})
-				),
-			pageFolderTree: (hideDefaultIndex = false) =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-
-					if (!status) {
-						const folderTree = yield* buildFolderTree;
-						const pages = yield* GET.pages(true, hideDefaultIndex);
-
-						for (const { data: page } of pages) {
-							if (page.parentFolder) {
-								yield* addPageToFolderTree(folderTree, page.parentFolder, {
-									id: page.id,
-									name: page.title,
-									page: true,
-									pageData: page,
-									children: [],
-								});
-							} else {
-								folderTree.push({
-									id: page.id,
-									name: page.title,
-									page: true,
-									pageData: page,
-									children: [],
-								});
-							}
-						}
-
-						return folderTreeReturn(folderTree);
-					}
-
-					const tree = pageFolderTree.get(PageFolderTreeMapID);
-
-					if (!tree || isCacheExpired(tree)) {
-						const newFolderTree = yield* buildFolderTree;
-						const pages = yield* GET.pages(true, hideDefaultIndex);
-
-						for (const { data: page } of pages) {
-							if (page.parentFolder) {
-								yield* addPageToFolderTree(newFolderTree, page.parentFolder, {
-									id: page.id,
-									name: page.title,
-									page: true,
-									pageData: page,
-									children: [],
-								});
-							} else {
-								newFolderTree.push({
-									id: page.id,
-									name: page.title,
-									page: true,
-									pageData: page,
-									children: [],
-								});
-							}
-						}
-
-						pageFolderTree.set(PageFolderTreeMapID, folderTreeReturn(newFolderTree));
-
-						return folderTreeReturn(newFolderTree);
-					}
-
-					return tree;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _clearLibSQLError('GET.pageFolderTree', cause),
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('GET.pageFolderTree', cause),
-					})
-				),
-			folderList: () =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-
-					if (!status) {
-						const folderList = yield* getAvailableFolders;
-
-						return folderListReturn(folderList);
-					}
-
-					const list = FolderList.get(FolderListMapID);
-
-					if (!list || isCacheExpired(list)) {
-						const folderList = yield* getAvailableFolders;
-
-						FolderList.set(FolderListMapID, folderListReturn(folderList));
-
-						return folderListReturn(folderList);
-					}
-
-					return list;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _clearLibSQLError('GET.folderList', cause),
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('GET.folderList', cause),
-					})
-				),
-			latestVersion: () =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-
-					if (!status) {
-						const version = yield* getVersionFromNPM.get(StudioCMSPkgId);
-						return versionReturn(version);
-					}
-
-					const latestVersion = version.get(VersionMapID);
-
-					if (!latestVersion || isCacheExpired(latestVersion, versionCacheLifetime)) {
-						const newVersion = yield* getVersionFromNPM.get(StudioCMSPkgId);
-
-						const latestVersion = versionReturn(newVersion);
-
-						version.set(VersionMapID, latestVersion);
-
-						return latestVersion;
-					}
-
-					return latestVersion;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _ClearUnknownError('GET.latestVersion', cause),
-					})
-				),
-			page: {
-				byId: _getPageById,
-				bySlug: _getPageBySlug,
-			},
-			folderPages: _folderPages,
-			packagePages: _getPackagesPages,
-			pages: _getAllPages,
-		};
-
-		const POST = {
-			databaseEntry: {
-				/**
-				 * Insert a new page into the database.
-				 *
-				 * @param pageData - The data to insert into the page data table.
-				 * @param pageContent - The data to insert into the page content table.
-				 * @returns A promise that resolves to the inserted page data and page content.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the page.
-				 */
-				pages: (pageData: tsPageDataSelect, pageContent: CombinedInsertContent) =>
-					Effect.gen(function* () {
-						const newContentID = pageData.id || crypto.randomUUID().toString();
-
-						const {
-							id = newContentID,
-							title,
-							slug,
-							description,
-							authorId = null,
-							package: packageName = 'studiocms',
-							contentLang = 'default',
-							heroImage = '',
-							showOnNav = false,
-							showAuthor = false,
-							showContributors = false,
-							categories = [],
-							tags = [],
-							contributorIds = [],
-							draft = false,
-							parentFolder = null,
-						} = pageData;
-
-						const stringified = {
-							categories: categories || [],
-							tags: tags || [],
-							contributorIds: contributorIds || [],
-						};
-
-						const contentData = {
-							id: crypto.randomUUID().toString(),
-							contentId: newContentID,
-							contentLang: contentLang,
-							content: pageContent.content || '',
-						};
-
-						const NOW = new Date();
-
-						const newPageData = yield* dbService.execute((db) =>
-							db
-								.insert(tsPageData)
-								.values({
-									// @ts-expect-error Drizzle... removed this from the type?
-									id,
-									title,
-									slug,
-									description,
-									authorId,
-									contentLang,
-									heroImage,
-									showAuthor,
-									showContributors,
-									showOnNav,
-									draft,
-									parentFolder,
-									package: packageName,
-									publishedAt: NOW,
-									updatedAt: NOW,
-									...stringified,
-								})
-								.returning({ id: tsPageData.id })
-						);
-
-						const newPageContent = yield* dbService.execute((db) =>
-							db.insert(tsPageContent).values(contentData).returning({ id: tsPageContent.id })
-						);
-
-						return {
-							pageData: newPageData,
-							pageContent: newPageContent,
-						} as addDatabaseEntryInsertPage;
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntry.pages', cause),
-						})
-					),
-				/**
-				 * Inserts new page content into the database.
-				 *
-				 * @param pageContent - The data to insert into the page content table.
-				 * @returns A promise that resolves to the inserted page content.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the page content.
-				 */
-				pageContent: dbService.makeQuery((ex, pageContent: tsPageContentSelect) =>
-					ex((db) =>
-						db
-							.insert(tsPageContent)
-							.values({
-								// @ts-expect-error Drizzle... removed this from the type?
-								id: pageContent.id || crypto.randomUUID().toString(),
-								contentId: pageContent.contentId,
-								contentLang: pageContent.contentLang || 'default',
-								content: pageContent.content || '',
-							})
-							.returning({ id: tsPageContent.id })
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntry.pageContent', cause),
-						})
-					)
-				),
-				/**
-				 * Inserts a new tag into the database.
-				 *
-				 * @param tag - The data to insert into the page data tags table.
-				 * @returns A promise that resolves to the inserted tag.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the tag.
-				 */
-				tags: (tag: tsPageDataTagsSelect) =>
-					Effect.gen(function* () {
-						const id = tag.id || (yield* generateRandomIDNumber(9));
-
-						return yield* dbService.execute((db) =>
-							db
-								.insert(tsPageDataTags)
-								.values({
-									// @ts-expect-error Drizzle... removed this from the type?
-									id,
-									name: tag.name,
-									description: tag.description,
-									slug: tag.slug,
-									meta: JSON.stringify(tag.meta),
-								})
-								.returning({ id: tsPageDataTags.id })
-						);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntry.tags', cause),
-						})
-					),
-				/**
-				 * Inserts a new category into the database.
-				 *
-				 * @param category - The data to insert into the page data categories table.
-				 * @returns A promise that resolves to the inserted category.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the category.
-				 */
-				categories: (category: tsPageDataCategoriesSelect) =>
-					Effect.gen(function* () {
-						const id = category.id || (yield* generateRandomIDNumber(9));
-
-						return yield* dbService.execute((db) =>
-							db
-								.insert(tsPageDataCategories)
-								.values({
-									// @ts-expect-error Drizzle... removed this from the type?
-									id,
-									name: category.name,
-									description: category.description,
-									slug: category.slug,
-									meta: JSON.stringify(category.meta),
-								})
-								.returning({ id: tsPageDataCategories.id })
-						);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntry.tags', cause),
-						})
-					),
-				/**
-				 * Inserts a new permission into the database.
-				 *
-				 * @param userId - The ID of the user to assign the rank to.
-				 * @param rank - The rank to assign to the user.
-				 * @returns A promise that resolves to the inserted permission.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the permission.
-				 */
-				permissions: (userId: string, rank: string) =>
-					Effect.gen(function* () {
-						const userExists = yield* dbService.execute((db) =>
-							db.select().from(tsPermissions).where(eq(tsPermissions.user, userId)).get()
-						);
-
-						if (userExists) {
-							return yield* Effect.fail(
-								yield* _clearLibSQLError(
-									'POST.databaseEntry.permissions',
-									'User already is already assigned a rank, please update the existing rank instead.'
-								)
-							);
-						}
-
-						return yield* dbService.execute((db) =>
-							db
-								.insert(tsPermissions)
-								.values({
-									user: userId,
-									rank,
-								})
-								.returning({ user: tsPermissions.user, rank: tsPermissions.rank })
-						);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntry.permissions', cause),
-						})
-					),
-				/**
-				 * Inserts a new diff tracking entry into the database.
-				 *
-				 * @param diff - The data to insert into the diff tracking table.
-				 * @returns A promise that resolves to the inserted diff tracking entry.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the diff tracking entry.
-				 */
-				diffTracking: dbService.makeQuery((ex, diff: tsDiffTrackingSelect) =>
-					ex((db) =>
-						db
-							.insert(tsDiffTracking)
-							.values({
-								// @ts-expect-error Drizzle... removed this from the type?
-								id: diff.id || crypto.randomUUID().toString(),
-								userId: diff.userId,
-								pageId: diff.pageId,
-								diff: diff.diff || '',
-								timestamp: diff.timestamp || new Date(),
-								pageContentStart: diff.pageContentStart,
-								pageMetaData: JSON.stringify(diff.pageMetaData || {}),
-							})
-							.returning()
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntry.diffTracking', cause),
-						})
-					)
-				),
-				/**
-				 * Inserts a new folder into the database.
-				 *
-				 * @param folder - The data to insert into the page folder structure table.
-				 * @returns A promise that resolves to the inserted folder.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the folder.
-				 */
-				folder: dbService.makeQuery((ex, folder: tsPageFolderSelect) =>
-					ex((db) =>
-						db
-							.insert(tsPageFolderStructure)
-							.values({
-								// @ts-expect-error Drizzle... removed this from the type?
-								id: folder.id || crypto.randomUUID().toString(),
-								name: folder.name,
-								parent: folder.parent || null,
-							})
-							.returning()
-							.get()
-					).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntry.folder', cause),
-						})
-					)
-				),
-			},
-			databaseEntries: {
-				/**
-				 * Inserts multiple tags into the database.
-				 *
-				 * @param data - The data to insert into the page data tags table.
-				 * @returns A promise that resolves to the inserted tags.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the tags.
-				 */
-				tags: (data: tsPageDataTagsSelect[]) =>
-					Effect.gen(function* () {
-						const entries: tsPageDataTagsSelect[] = [];
-
-						for (const item of data) {
-							const id = item.id || (yield* generateRandomIDNumber(9));
-							entries.push({
-								id,
-								name: item.name,
-								slug: item.slug,
-								description: item.description,
-								meta: JSON.stringify(item.meta),
-							});
-						}
-
-						return yield* dbService.execute((db) =>
-							db.insert(tsPageDataTags).values(entries).returning({ id: tsPageDataTags.id })
-						);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntries.tags', cause),
-						})
-					),
-				/**
-				 * Inserts multiple categories into the database.
-				 *
-				 * @param data - The data to insert into the page data categories table.
-				 * @returns A promise that resolves to the inserted categories.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the categories.
-				 */
-				categories: (data: tsPageDataCategoriesSelect[]) =>
-					Effect.gen(function* () {
-						const entries: tsPageDataCategoriesSelect[] = [];
-
-						for (const item of data) {
-							const id = item.id || (yield* generateRandomIDNumber(9));
-							entries.push({
-								id,
-								name: item.name,
-								slug: item.slug,
-								description: item.description,
-								meta: JSON.stringify(item.meta),
-								parent: null,
-							});
-						}
-
-						return yield* dbService.execute((db) =>
-							db
-								.insert(tsPageDataCategories)
-								.values(entries)
-								.returning({ id: tsPageDataCategories.id })
-						);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntries.categories', cause),
-						})
-					),
-				/**
-				 * Inserts multiple permissions into the database.
-				 *
-				 * @param data - The data to insert into the permissions table.
-				 * @returns A promise that resolves to the inserted permissions.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the permissions.
-				 */
-				permissions: (data: tsPermissionsInsert[]) =>
-					Effect.gen(function* () {
-						const currentPermittedUsers = yield* dbService.execute((db) =>
-							db.select().from(tsPermissions)
-						);
-
-						for (const permission of data) {
-							const userAlreadyExists = currentPermittedUsers.find(
-								(user) => user.user === permission.user
-							);
-
-							if (userAlreadyExists) {
-								return yield* Effect.fail(
-									new SDKCoreError({
-										type: 'LibSQLDatabaseError',
-										cause: new StudioCMS_SDK_Error(
-											`User with ID ${permission.user} already has a rank assigned. Please update the existing rank instead.`
-										),
-									})
-								);
-							}
-						}
-
-						return yield* dbService.execute((db) =>
-							db
-								.insert(tsPermissions)
-								.values(
-									data.map((permission) => {
-										return {
-											user: permission.user,
-											rank: permission.rank,
-										};
-									})
-								)
-								.returning({ user: tsPermissions.user, rank: tsPermissions.rank })
-						);
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('POST.databaseEntries.permissions', cause),
-						})
-					),
-				/**
-				 * Inserts multiple pages into the database.
-				 *
-				 * @param pages - The data to insert into the page data and page content tables.
-				 * @returns A promise that resolves to the inserted pages.
-				 * @throws {StudioCMS_SDK_Error} If an error occurs while inserting the pages.
-				 */
-				pages: (pages: MultiPageInsert) =>
-					Effect.gen(function* () {
-						for (const { pageData, pageContent } of pages) {
-							yield* POST.databaseEntry.pages(pageData, pageContent);
-						}
-					}),
-			},
-			folder: (data: tsPageFolderSelect) =>
-				Effect.gen(function* () {
-					const newEntry = yield* POST.databaseEntry.folder(data);
-
-					yield* CLEAR.folderList();
-					yield* CLEAR.folderTree();
-
-					yield* UPDATE.folderList;
-					yield* UPDATE.folderTree;
-
-					return newEntry;
-				}),
-			page: (data: {
-				pageData: tsPageDataSelect;
-				pageContent: CombinedInsertContent;
-			}) =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-
-					if (!status) {
-						const newPage = yield* POST.databaseEntry.pages(data.pageData, data.pageContent);
-
-						const { data: toReturn } = yield* GET.page.byId(newPage.pageData[0].id);
-
-						return pageDataReturn(toReturn);
-					}
-
-					const newPage = yield* POST.databaseEntry.pages(data.pageData, data.pageContent);
-
-					const { data: toReturn } = yield* GET.page.byId(newPage.pageData[0].id);
-
-					pages.set(toReturn.id, pageDataReturn(toReturn));
-					yield* CLEAR.folderList();
-					yield* CLEAR.folderTree();
-
-					return pageDataReturn(toReturn);
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _ClearUnknownError('POST.page', cause),
-					})
-				),
-		};
-
-		const UPDATE = {
-			/**
-			 * Updates a page content in the database.
-			 *
-			 * @param data - The data to update in the page content table.
-			 * @returns A promise that resolves to the updated page content.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while updating the page content.
-			 */
-			pageContent: dbService.makeQuery((ex, data: tsPageContentSelect) =>
-				ex((db) =>
-					db.update(tsPageContent).set(data).where(eq(tsPageContent.id, data.id)).returning().get()
-				).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('UPDATE.pageContent', cause),
-					})
-				)
-			),
-			/**
-			 * Updates a tag in the database.
-			 *
-			 * @param data - The data to update in the page data tags table.
-			 * @returns A promise that resolves to the updated tag.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while updating the tag.
-			 */
-			tags: dbService.makeQuery((ex, data: tsPageDataTagsSelect) =>
-				ex((db) =>
-					db
-						.update(tsPageDataTags)
-						.set(data)
-						.where(eq(tsPageDataTags.id, data.id))
-						.returning()
-						.get()
-				).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('UPDATE.tags', cause),
-					})
-				)
-			),
-			/**
-			 * Updates a category in the database.
-			 *
-			 * @param data - The data to update in the page data categories table.
-			 * @returns A promise that resolves to the updated category.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while updating the category.
-			 */
-			categories: dbService.makeQuery((ex, data: tsPageDataCategoriesSelect) =>
-				ex((db) =>
-					db
-						.update(tsPageDataCategories)
-						.set(data)
-						.where(eq(tsPageDataCategories.id, data.id))
-						.returning()
-						.get()
-				).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('UPDATE.categories', cause),
-					})
-				)
-			),
-			/**
-			 * Updates a permission in the database.
-			 *
-			 * @param data - The data to update in the permissions table.
-			 * @returns A promise that resolves to the updated permission.
-			 * @throws {StudioCMS_SDK_Error} If an error occurs while updating the permission.
-			 */
-			permissions: dbService.makeQuery((ex, data: tsPermissionsSelect) =>
-				ex((db) =>
-					db
-						.update(tsPermissions)
-						.set(data)
-						.where(eq(tsPermissions.user, data.user))
-						.returning()
-						.get()
-				).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('UPDATE.permissions', cause),
-					})
-				)
-			),
-			folderTree: Effect.gen(function* () {
-				const status = yield* isCacheEnabled;
-				yield* CLEAR.folderTree();
-				const newFolderTree = yield* buildFolderTree;
-				if (status) {
-					folderTree.set(FolderTreeMapID, folderTreeReturn(newFolderTree));
-				}
-			}).pipe(
-				Effect.catchTags({
-					'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-						_clearLibSQLError('UPDATE.folderTree', cause),
-				})
-			),
-			folderList: Effect.gen(function* () {
-				const status = yield* isCacheEnabled;
-				yield* CLEAR.folderList();
-				const folderList = yield* getAvailableFolders;
-				if (status) {
-					FolderList.set(FolderListMapID, folderListReturn(folderList));
-				}
-			}).pipe(
-				Effect.catchTags({
-					'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-						_clearLibSQLError('UPDATE.folderList', cause),
-				})
-			),
-			folder: (data: tsPageFolderSelect) =>
-				Effect.gen(function* () {
-					const updated = yield* dbService.execute((db) =>
-						db
-							.update(tsPageFolderStructure)
-							.set(data)
-							.where(eq(tsPageFolderStructure.id, data.id))
-							.returning()
-							.get()
-					);
-
-					yield* UPDATE.folderList;
-					yield* UPDATE.folderTree;
-
-					return updated;
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('UPDATE.folder', cause),
-					})
-				),
-			latestVersion: () =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-					const latestVersion = yield* getVersionFromNPM.get(StudioCMSPkgId);
-
-					const newVersion = versionReturn(latestVersion);
-
-					if (!status) return newVersion;
-
-					version.set(VersionMapID, newVersion);
-
-					return newVersion;
-				}).pipe(
-					Effect.catchTags({
-						UnknownException: (cause) => _ClearUnknownError('UPDATE.latestVersion', cause),
-					})
-				),
-			siteConfig: (data: SiteConfig) =>
-				Effect.gen(function* () {
-					const status = yield* isCacheEnabled;
-					const newSiteConfig = yield* dbService.execute((db) =>
-						db
-							.update(tsSiteConfig)
-							.set(data)
-							.where(eq(tsSiteConfig.id, CMSSiteConfigId))
-							.returning()
-							.get()
-					);
-
-					const returnConfig = siteConfigReturn(newSiteConfig);
-
-					if (!status) return returnConfig;
-
-					siteConfig.set(SiteConfigMapID, returnConfig);
-
-					return returnConfig;
-				}).pipe(
-					Effect.catchTags({
-						'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-							_clearLibSQLError('UPDATE.siteConfig', cause),
-						UnknownException: (cause) => _ClearUnknownError('UPDATE.siteConfig', cause),
-					})
-				),
-			page: {
-				byId: (
-					id: string,
-					data: {
-						pageData: tsPageDataSelect;
-						pageContent: tsPageContentSelect;
-					}
-				) => {
-					const updatePage = dbService.makeQuery((ex, data: tsPageDataSelect) =>
-						ex((db) =>
-							db.update(tsPageData).set(data).where(eq(tsPageData.id, data.id)).returning().get()
-						)
-					);
-
-					return Effect.gen(function* () {
-						const status = yield* isCacheEnabled;
-
-						yield* updatePage(data.pageData);
-						yield* UPDATE.pageContent(data.pageContent);
-
-						const { data: updatedData } = yield* GET.page.byId(id);
-						const returnData = pageDataReturn(updatedData);
-
-						if (!status) {
-							return returnData;
-						}
-
-						pages.set(id, returnData);
-						yield* CLEAR.folderList();
-						yield* CLEAR.folderTree();
-
-						return returnData;
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('UPDATE.page.byId', cause),
-							UnknownException: (cause) => _ClearUnknownError('UPDATE.page.byId', cause),
-						})
-					);
-				},
-				bySlug: (
-					slug: string,
-					data: {
-						pageData: tsPageDataSelect;
-						pageContent: tsPageContentSelect;
-					}
-				) => {
-					const updatePage = dbService.makeQuery((ex, data: tsPageDataSelect) =>
-						ex((db) =>
-							db.update(tsPageData).set(data).where(eq(tsPageData.id, data.id)).returning().get()
-						)
-					);
-
-					return Effect.gen(function* () {
-						const status = yield* isCacheEnabled;
-
-						if (!status) {
-							yield* updatePage(data.pageData);
-							yield* UPDATE.pageContent(data.pageContent);
-
-							const { data: updatedData } = yield* GET.page.bySlug(slug);
-
-							return pageDataReturn(updatedData);
-						}
-
-						const cachedPage = Array.from(pages.values()).find((page) => page.data.slug === slug);
-
-						if (!cachedPage) {
-							return yield* Effect.fail(
-								yield* new SDKCoreError({
-									type: 'UNKNOWN',
-									cause: new StudioCMS_SDK_Error('Page not found in cache'),
-								})
-							);
-						}
-
-						yield* updatePage(data.pageData);
-						yield* UPDATE.pageContent(data.pageContent);
-
-						const { data: updatedData } = yield* GET.page.bySlug(slug);
-
-						const returnData = pageDataReturn(updatedData);
-
-						pages.set(updatedData.id, returnData);
-
-						yield* CLEAR.folderList();
-						yield* CLEAR.folderTree();
-
-						return returnData;
-					}).pipe(
-						Effect.catchTags({
-							'studiocms/sdk/effect/db/LibSQLDatabaseError': (cause) =>
-								_clearLibSQLError('UPDATE.page.byId', cause),
-							UnknownException: (cause) => _ClearUnknownError('UPDATE.page.byId', cause),
-						})
-					);
-				},
-			},
-		};
-
 		return {
 			db,
 			dbService,
@@ -3382,38 +1185,20 @@ export class SDKCore extends Effect.Service<SDKCore>()('studiocms/sdk/SDKCore', 
 			collectTags,
 			collectPageData,
 			collectUserData,
+			CLEAR, // migrated
+			DELETE, // migrated
+			REST_API, // migrated
+			POST, // migrated
+			UPDATE, // migrated
+			GET, // migrated
+			INIT,
+			AUTH,
 			resetTokenBucket,
 			diffTracking,
 			notificationSettings,
-			CLEAR,
-			DELETE,
-			INIT,
-			AUTH,
-			REST_API,
-			GET,
-			POST,
-			UPDATE,
 		};
 	}),
-	dependencies: [
-		AstroDB.Default,
-		SDKCore_FolderTree.Default,
-		SDKCore_Generators.Default,
-		SDKCore_Parsers.Default,
-		SDKCore_Users.Default,
-		SDKCore_Collectors.Default,
-		GetVersionFromNPM.Default,
-	],
 }) {
 	static Provide = Effect.provide(this.Default);
+	static Cache = CacheContext.makeProvide({ pages, FolderList, folderTree, pageFolderTree, siteConfig, version })
 }
-
-/**
- * Test Effect for testing how SDKCore is shaped
- */
-// const testProgram = Effect.gen(function* () {
-// 	const sdkC = yield* SDKCore;
-// 	return sdkC;
-// }).pipe(Effect.provide(SDKCore.Default));
-
-// const program = await Effect.runPromise(testProgram);
